Removed extraneous files and fixed scraper bugs.
Updated randomCollectionIndex() to say the collection can't be empty.

Added isEmpty() check for link fetching.

Changed the scraper to not fetch titles and changed method in ArrayUtil.

Changed the scraper to not fetch titles.
Changed prependAndAppendStringToList() to prependStringToList()
Created a javadoc for the new method.
Changed variable types to be List instead of ArrayList.

Fixed infinite execution when sources can't be retrieved

Added blank string returns in getNextExcerptInSentences

Added blank string return in getNextExcerptInSentences when
approximateCharacterCount is 0

Added blank string return in getNextExcerptInSentences when
fileSentences is null instead of returning null.

Added a message to display when texts are being regenerated.

Added an issue message when fetched text is blank.

Made texts properly wrap when end of text from file is reached

Fixed La Nación scraper after webpage redesign.